+++ src/Compute/Compute/Usage/NewAzureSshKey.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Compute.Automation.Models;
@@ -43,6 +45,16 @@ namespace Microsoft.Azure.Commands.Compute.Automation
         [ValidateSet("Ed25519", "RSA")]
         public string SshKeyType { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true)]
+        [LocationCompleter("Microsoft.Compute/SshPublicKeys")]
+        public string Location { get; set; }
+
+        [Parameter(
+            Mandatory = false)]
+        public Hashtable Tag { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -53,8 +65,21 @@ namespace Microsoft.Azure.Commands.Compute.Automation
                 string sshKeyType = this.SshKeyType;
                 SshPublicKeyResource result;
                 SshPublicKeyResource sshkey = new SshPublicKeyResource();
-                ResourceGroup rg = ArmClient.ResourceGroups.Get(resourceGroupName);
-                sshkey.Location = rg.Location;
+                
+                if (this.IsParameterBound(c => c.Location))
+                {
+                    sshkey.Location = this.Location;
+                }
+                else
+                {
+                    ResourceGroup rg = ArmClient.ResourceGroups.Get(resourceGroupName);
+                    sshkey.Location = rg.Location;
+                }
+
+                if (this.MyInvocation.BoundParameters.ContainsKey("Tag"))
+                {
+                    sshkey.Tags = this.Tag.Cast<DictionaryEntry>().ToDictionary(ht => (string)ht.Key, ht => (string)ht.Value);
+                }
 
 
                 if (this.IsParameterBound(c => c.PublicKey))